Revert "Почти сделал связывание"
This reverts commit bcd0e170271ad06d348f5c4c0e2d8ec8156c90a6.
